+++ contrib/pg_upgrade/Makefile
@@ -10,7 +10,9 @@ OBJS = check.o controldata.o dump.o exec.o file.o function.o info.o \
        option.o page.o pg_upgrade.o relfilenode.o server.o \
        tablespace.o util.o version.o version_old_8_3.o $(WIN32RES)
 
-OBJS += aotable.o gpdb4_heap_convert.o version_old_gpdb4.o oid_dump.o
+# Source files specific to Greenplum
+OBJS += aotable.o gpdb4_heap_convert.o version_gp.o oid_dump.o \
+        check_gp.o file_gp.o reporting.o
 
 PG_CPPFLAGS  = -DFRONTEND -DDLSUFFIX=\"$(DLSUFFIX)\" -I$(srcdir) -I$(libpq_srcdir)
 PG_LIBS = $(libpq_pgport)
