+++ contrib/pg_upgrade/check_gp.c
@@ -1,0 +1,347 @@
+/*
+ *	check_gp.c
+ *
+ *	Greenplum specific server checks and output routines
+ *
+ *	Any compatibility checks which are version dependent (testing for issues in
+ *	specific versions of Greenplum) should be placed in their respective
+ *	version_old_gpdb{MAJORVERSION}.c file.  The checks in this file supplement
+ *	the checks present in check.c, which is the upstream file for performing
+ *	checks against a PostgreSQL cluster.
+ *
+ *	Copyright (c) 2010, PostgreSQL Global Development Group
+ *	Copyright (c) 2017-Present Pivotal Software, Inc
+ */
+
+#include "pg_upgrade.h"
+
+static void check_external_partition(void);
+static void check_covering_aoindex(void);
+static void check_partition_indexes(void);
+
+/*
+ *	check_greenplum
+ *
+ *	Rather than exporting all checks, we export a single API function which in
+ *	turn is responsible for running Greenplum checks. This function should be
+ *	executed after all PostgreSQL checks. The order of the checks should not
+ *	matter.
+ */
+void
+check_greenplum(void)
+{
+	/*
+	 * Upgrading within a major version is a handy feature of pg_upgrade, but
+	 * we don't allow it for within 4.3.x clusters, 4.3.x can only be an old
+	 * version to be upgraded from.
+	 */
+	if (GET_MAJOR_VERSION(old_cluster.major_version) <= 802 &&
+		GET_MAJOR_VERSION(new_cluster.major_version) <= 802)
+	{
+		pg_log(PG_FATAL,
+			   "old and new cluster cannot both be Greenplum 4.3.x installations\n");
+	}
+
+	check_external_partition();
+	check_covering_aoindex();
+	check_partition_indexes();
+}
+
+/*
+ *	check_external_partition
+ *
+ *	External tables cannot be included in the partitioning hierarchy during the
+ *	initial definition with CREATE TABLE, they must be defined separately and
+ *	injected via ALTER TABLE EXCHANGE. The partitioning system catalogs are
+ *	however not replicated onto the segments which means ALTER TABLE EXCHANGE
+ *	is prohibited in utility mode. This means that pg_upgrade cannot upgrade a
+ *	cluster containing external partitions, they must be handled manually
+ *	before/after the upgrade.
+ *
+ *	Check for the existence of external partitions and refuse the upgrade if
+ *	found.
+ */
+static void
+check_external_partition(void)
+{
+	char		query[QUERY_ALLOC];
+	char		output_path[MAXPGPATH];
+	FILE	   *script = NULL;
+	bool		found = false;
+	int			dbnum;
+
+	prep_status("Checking for external tables used in partitioning");
+
+	snprintf(output_path, sizeof(output_path), "%s/external_partitions.txt",
+			 os_info.cwd);
+	/*
+	 * We need to query the inheritance catalog rather than the partitioning
+	 * catalogs since they are not available on the segments.
+	 */
+	snprintf(query, sizeof(query),
+			 "SELECT cc.relname, c.relname AS partname, c.relnamespace "
+			 "FROM   pg_inherits i "
+			 "       JOIN pg_class c ON (i.inhrelid = c.oid AND c.relstorage = '%c') "
+			 "       JOIN pg_class cc ON (i.inhparent = cc.oid);",
+			 RELSTORAGE_EXTERNAL);
+
+	for (dbnum = 0; dbnum < old_cluster.dbarr.ndbs; dbnum++)
+	{
+		PGresult   *res;
+		int			ntups;
+		int			rowno;
+		DbInfo	   *active_db = &old_cluster.dbarr.dbs[dbnum];
+		PGconn	   *conn;
+
+		conn = connectToServer(&old_cluster, active_db->db_name);
+		res = executeQueryOrDie(conn, query);
+
+		ntups = PQntuples(res);
+
+		if (ntups > 0)
+		{
+			found = true;
+
+			if (script == NULL && (script = fopen(output_path, "w")) == NULL)
+				pg_log(PG_FATAL, "Could not create necessary file:  %s\n",
+					   output_path);
+
+			for (rowno = 0; rowno < ntups; rowno++)
+			{
+				fprintf(script, "External partition \"%s\" in relation \"%s\"\n",
+						PQgetvalue(res, rowno, PQfnumber(res, "partname")),
+						PQgetvalue(res, rowno, PQfnumber(res, "relname")));
+			}
+		}
+
+		PQclear(res);
+		PQfinish(conn);
+	}
+	if (found)
+	{
+		fclose(script);
+		pg_log(PG_REPORT, "fatal\n");
+		pg_log(PG_FATAL,
+			   "| Your installation contains partitioned tables with external\n"
+			   "| tables as partitions.  These partitions need to be removed\n"
+			   "| from the partition hierarchy before the upgrade.  A list of\n"
+			   "| external partitions to remove is in the file:\n"
+			   "| \t%s\n\n", output_path);
+	}
+	else
+	{
+		check_ok();
+	}
+}
+
+/*
+ *	check_covering_aoindex
+ *
+ *	A partitioned AO table which had an index created on the parent relation,
+ *	and an AO partition exchanged into the hierarchy without any indexes will
+ *	break upgrades due to the way pg_dump generates DDL.
+ *
+ *	create table t (a integer, b text, c integer)
+ *		with (appendonly=true)
+ *		distributed by (a)
+ *		partition by range(c) (start(1) end(3) every(1));
+ *	create index t_idx on t (b);
+ *
+ *	At this point, the t_idx index has created AO blockdir relations for all
+ *	partitions. We now exchange a new table into the hierarchy which has no
+ *	index defined:
+ *
+ *	create table t_exch (a integer, b text, c integer)
+ *		with (appendonly=true)
+ *		distributed by (a);
+ *	alter table t exchange partition for (rank(1)) with table t_exch;
+ *
+ *	The partition which was swapped into the hierarchy with EXCHANGE does not
+ *	have any indexes and thus no AO blockdir relation. This is in itself not
+ *	a problem, but when pg_dump generates DDL for the above situation it will
+ *	create the index in such a way that it covers the entire hierarchy, as in
+ *	its original state. The below snippet illustrates the dumped DDL:
+ *
+ *	create table t ( ... )
+ *		...
+ *		partition by (... );
+ *	create index t_idx on t ( ... );
+ *
+ *	This creates a problem for the Oid synchronization in pg_upgrade since it
+ *	expects to find a preassigned Oid for the AO blockdir relations for each
+ *	partition. A longer term solution would be to generate DDL in pg_dump which
+ *	creates the current state, but for the time being we disallow upgrades on
+ *	cluster which exhibits this.
+ */
+static void
+check_covering_aoindex(void)
+{
+	char			query[QUERY_ALLOC];
+	char			output_path[MAXPGPATH];
+	FILE		   *script = NULL;
+	bool			found = false;
+	int				dbnum;
+
+	prep_status("Checking for non-covering indexes on partitioned AO tables");
+
+	snprintf(output_path, sizeof(output_path), "%s/mismatched_aopartition_indexes.txt",
+			 os_info.cwd);
+
+	snprintf(query, sizeof(query),
+			 "SELECT DISTINCT ao.relid, inh.inhrelid "
+			 "FROM   pg_catalog.pg_appendonly ao "
+			 "       JOIN pg_catalog.pg_inherits inh "
+			 "         ON (inh.inhparent = ao.relid) "
+			 "       JOIN pg_catalog.pg_appendonly aop "
+			 "         ON (inh.inhrelid = aop.relid AND aop.blkdirrelid = 0) "
+			 "       JOIN pg_catalog.pg_index i "
+			 "         ON (i.indrelid = ao.relid) "
+			 "WHERE  ao.blkdirrelid <> 0;");
+
+	for (dbnum = 0; dbnum < old_cluster.dbarr.ndbs; dbnum++)
+	{
+		PGresult   *res;
+		PGconn	   *conn;
+		int			ntups;
+		int			rowno;
+		DbInfo	   *active_db = &old_cluster.dbarr.dbs[dbnum];
+
+		conn = connectToServer(&old_cluster, active_db->db_name);
+		res = executeQueryOrDie(conn, query);
+
+		ntups = PQntuples(res);
+
+		if (ntups > 0)
+		{
+			found = true;
+
+			if (script == NULL && (script = fopen(output_path, "w")) == NULL)
+				pg_log(PG_FATAL, "Could not create necessary file:  %s\n",
+					   output_path);
+
+			for (rowno = 0; rowno < ntups; rowno++)
+			{
+				fprintf(script, "Mismatched index on partition %s in relation %s\n",
+						PQgetvalue(res, rowno, PQfnumber(res, "inhrelid")),
+						PQgetvalue(res, rowno, PQfnumber(res, "relid")));
+			}
+		}
+
+		PQclear(res);
+		PQfinish(conn);
+	}
+
+	if (found)
+	{
+		fclose(script);
+		pg_log(PG_REPORT, "fatal\n");
+		pg_log(PG_FATAL,
+			   "| Your installation contains partitioned append-only tables\n"
+			   "| with an index defined on the partition parent which isn't\n"
+			   "| present on all partition members.  These indexes must be\n"
+			   "| dropped before the upgrade.  A list of relations, and the\n"
+			   "| partitions in question is in the file:\n"
+			   "| \t%s\n\n", output_path);
+
+	}
+	else
+	{
+		check_ok();
+	}
+}
+
+/*
+ *	check_partition_indexes
+ *
+ *	There are numerous pitfalls surrounding indexes on partition hierarchies,
+ *	so rather than trying to cover all the cornercases we disallow indexes on
+ *	partitioned tables altogether during the upgrade.  Since we in any case
+ *	invalidate the indexes forcing a REINDEX, there is little to be gained by
+ *	handling them for the end-user.
+ */
+static void
+check_partition_indexes(void)
+{
+	int				dbnum;
+	FILE		   *script = NULL;
+	bool			found = false;
+	char			output_path[MAXPGPATH];
+
+	prep_status("Checking for indexes on partitioned tables");
+
+	snprintf(output_path, sizeof(output_path), "%s/partitioned_tables_indexes.txt",
+			 os_info.cwd);
+
+	for (dbnum = 0; dbnum < old_cluster.dbarr.ndbs; dbnum++)
+	{
+		PGresult   *res;
+		bool		db_used = false;
+		int			ntups;
+		int			rowno;
+		int			i_nspname;
+		int			i_relname;
+		int			i_indexes;
+		DbInfo	   *active_db = &old_cluster.dbarr.dbs[dbnum];
+		PGconn	   *conn = connectToServer(&old_cluster, active_db->db_name);
+
+		res = executeQueryOrDie(conn,
+								"WITH partitions AS ("
+								"    SELECT DISTINCT n.nspname, "
+								"           c.relname "
+								"    FROM pg_catalog.pg_partition p "
+								"         JOIN pg_catalog.pg_class c ON (p.parrelid = c.oid) "
+								"         JOIN pg_catalog.pg_namespace n ON (n.oid = c.relnamespace) "
+								"    UNION "
+								"    SELECT n.nspname, "
+								"           partitiontablename AS relname "
+								"    FROM pg_catalog.pg_partitions p "
+								"         JOIN pg_catalog.pg_class c ON (p.partitiontablename = c.relname) "
+								"         JOIN pg_catalog.pg_namespace n ON (n.oid = c.relnamespace) "
+								") "
+								"SELECT nspname, "
+								"       relname, "
+								"       count(indexname) AS indexes "
+								"FROM partitions "
+								"     JOIN pg_catalog.pg_indexes ON (relname = tablename AND "
+								"                                    nspname = schemaname) "
+								"GROUP BY nspname, relname "
+								"ORDER BY relname");
+
+		ntups = PQntuples(res);
+		i_nspname = PQfnumber(res, "nspname");
+		i_relname = PQfnumber(res, "relname");
+		i_indexes = PQfnumber(res, "indexes");
+		for (rowno = 0; rowno < ntups; rowno++)
+		{
+			found = true;
+			if (script == NULL && (script = fopen(output_path, "w")) == NULL)
+				pg_log(PG_FATAL, "Could not create necessary file:  %s\n", output_path);
+			if (!db_used)
+			{
+				fprintf(script, "Database:  %s\n", active_db->db_name);
+				db_used = true;
+			}
+			fprintf(script, "  %s.%s has %s index(es)\n",
+					PQgetvalue(res, rowno, i_nspname),
+					PQgetvalue(res, rowno, i_relname),
+					PQgetvalue(res, rowno, i_indexes));
+		}
+
+		PQclear(res);
+		PQfinish(conn);
+	}
+
+	if (found)
+	{
+		fclose(script);
+		pg_log(PG_REPORT, "fatal\n");
+		pg_log(PG_FATAL,
+			   "| Your installation contains partitioned tables with\n"
+			   "| indexes defined on them.  Indexes on partition parents,\n"
+			   "| as well as children, must be dropped before upgrade.\n"
+			   "| A list of the problem tables is in the file:\n"
+			   "| \t%s\n\n", output_path);
+	}
+	else
+		check_ok();
+}
