+++ contrib/pg_upgrade/file.c
@@ -11,9 +11,6 @@
 
 #include <fcntl.h>
 
-#include "storage/bufpage.h"
-#include "storage/checksum.h"
-#include "storage/checksum_impl.h"
 
 
 #ifndef WIN32
@@ -108,92 +105,6 @@ copyAndUpdateFile(pageCnvCtx *pageConverter,
 	}
 }
 
-/*
- * rewriteHeapPageWithChecksum
- *
- * Copies a relation file from src to dst and sets the data checksum in the
- * page headers in the process. We are not using a pageConverter, even though
- * that would make sense, since pageConverter are deprecated and removed in
- * upstream and would give us merge headaches.
- */
-void
-rewriteHeapPageChecksum(const char *fromfile, const char *tofile,
-						const char *schemaName, const char *relName)
-{
-	int			src_fd;
-	int			dst_fd;
-	int			blkno;
-	int			bytesRead;
-	int			totalBytesRead;
-	char	   *buf;
-	ssize_t		writesize;
-	struct stat statbuf;
-
-	/*
-	 * transfer_relfile() should never call us unless requested by the data
-	 * checksum option but better doublecheck before we start rewriting data.
-	 */
-	if (user_opts.checksum_mode == CHECKSUM_NONE)
-		pg_log(PG_FATAL, "error, incorrect checksum configuration detected.\n");
-
-	if ((src_fd = open(fromfile, O_RDONLY | PG_BINARY, 0)) < 0)
-		pg_log(PG_FATAL,
-			   "error while rewriting relation \"%s.%s\": could not open file \"%s\": %s\n",
-			   schemaName, relName, fromfile, strerror(errno));
-
-	if (fstat(src_fd, &statbuf) != 0)
-		pg_log(PG_FATAL,
-			   "error while rewriting relation \"%s.%s\": could not stat file \"%s\": %s\n",
-			   schemaName, relName, fromfile, strerror(errno));
-
-	if ((dst_fd = open(tofile, O_RDWR | O_CREAT | O_EXCL | PG_BINARY, S_IRUSR | S_IWUSR)) < 0)
-		pg_log(PG_FATAL,
-			   "error while rewriting relation \"%s.%s\": could not create file \"%s\": %s\n",
-			   schemaName, relName, tofile, strerror(errno));
-
-	blkno = 0;
-	totalBytesRead = 0;
-	buf = (char *) pg_malloc(BLCKSZ);
-
-	while ((bytesRead = read(src_fd, buf, BLCKSZ)) == BLCKSZ)
-	{
-		Size		page_size;
-
-		page_size = PageGetPageSize((PageHeader) buf);
-
-		if (!PageSizeIsValid(page_size) && page_size != 0)
-			pg_log(PG_FATAL,
-				   "error while rewriting relation \"%s.%s\": invalid page size detected (%zd)\n",
-				   schemaName, relName, page_size);
-
-		if (!PageIsNew(buf))
-		{
-			if (user_opts.checksum_mode == CHECKSUM_ADD)
-				((PageHeader) buf)->pd_checksum = pg_checksum_page(buf, blkno);
-			else
-				memset(&(((PageHeader) buf)->pd_checksum), 0, sizeof(uint16));
-		}
-
-		writesize = write(dst_fd, buf, BLCKSZ);
-
-		if (writesize != BLCKSZ)
-			pg_log(PG_FATAL, "error when rewriting relation \"%s.%s\": %s",
-				   schemaName, relName, strerror(errno));
-
-		blkno++;
-		totalBytesRead += BLCKSZ;
-	}
-
-	if (totalBytesRead != statbuf.st_size)
-		pg_log(PG_FATAL,
-			   "error when rewriting relation \"%s.%s\": torn read on file \"%s\"%c %s\n",
-			   schemaName, relName, fromfile,
-			   (errno != 0 ? ':' : ' '), strerror(errno));
-
-	pg_free(buf);
-	close(dst_fd);
-	close(src_fd);
-}
 
 /*
  * linkAndUpdateFile()
