+++ contrib/pg_upgrade/version.c
@@ -89,125 +89,3 @@ new_9_0_populate_pg_largeobject_metadata(ClusterInfo *cluster, bool check_mode)
 	else
 		check_ok();
 }
-
-
-/*
- * new_gpdb5_0_invalidate_indexes()
- *	new >= GPDB 5.0, old <= GPDB 4.3
- *
- * GPDB 5.0 follows the PostgreSQL 8.3 page format, while GPDB 4.3 used
- * the 8.2 format. A new field was added to the page header, so we need
- * mark all indexes as invalid.
- */
-void
-new_gpdb5_0_invalidate_indexes(bool check_mode)
-{
-	int			dbnum;
-	FILE	   *script = NULL;
-	char		output_path[MAXPGPATH];
-
-	prep_status("Invalidating indexes in new cluster");
-
-	snprintf(output_path, sizeof(output_path), "%s/reindex_all.sql",
-			 os_info.cwd);
-
-	if (!check_mode)
-	{
-		if ((script = fopen(output_path, "w")) == NULL)
-			pg_log(PG_FATAL, "Could not create necessary file:  %s\n",
-				   output_path);
-	}
-
-	for (dbnum = 0; dbnum < old_cluster.dbarr.ndbs; dbnum++)
-	{
-		DbInfo	   *olddb = &old_cluster.dbarr.dbs[dbnum];
-		PGconn	   *conn = connectToServer(&new_cluster, olddb->db_name);
-		char		query[QUERY_ALLOC];
-
-		/*
-		 * GPDB doesn't allow hacking the catalogs without setting
-		 * allow_system_table_mods first.
-		 */
-		PQclear(executeQueryOrDie(conn, "set allow_system_table_mods='dml'"));
-
-		/*
-		 * check_mode doesn't do much interesting for this but at least
-		 * we'll know we are allowed to change allow_system_table_mods
-		 * which is required
-		 */
-		if (!check_mode)
-		{
-			snprintf(query, sizeof(query),
-					 "UPDATE pg_index SET indisvalid = false WHERE indexrelid >= %u",
-					 FirstNormalObjectId);
-			PQclear(executeQueryOrDie(conn, query));
-
-			fprintf(script, "\\connect %s\n",
-					quote_identifier(olddb->db_name));
-			fprintf(script, "REINDEX DATABASE %s;\n",
-					quote_identifier(olddb->db_name));
-		}
-		PQfinish(conn);
-	}
-
-	if (!check_mode)
-		fclose(script);
-	report_status(PG_WARNING, "warning");
-	if (check_mode)
-		pg_log(PG_WARNING, "\n"
-			   "| All indexes have different internal formats\n"
-			   "| between your old and new clusters so they must\n"
-			   "| be reindexed with the REINDEX command. After\n"
-			   "| migration, you will be given REINDEX instructions.\n\n");
-	else
-		pg_log(PG_WARNING, "\n"
-			   "| All indexes have different internal formats\n"
-			   "| between your old and new clusters so they must\n"
-			   "| be reindexed with the REINDEX command.\n"
-			   "| The file:\n"
-			   "| \t%s\n"
-			   "| when executed by psql by the database super-user\n"
-			   "| will recreate all invalid indexes.\n\n",
-			   output_path);
-}
-
-/*
- * new_gpdb_invalidate_bitmap_indexes()
- *
- * TODO: We are currently missing the support to migrate over bitmap indexes.
- * Hence, mark all bitmap indexes as invalid.
- */
-void
-new_gpdb_invalidate_bitmap_indexes(bool check_mode)
-{
-	int			dbnum;
-
-	prep_status("Invalidating indexes in new cluster");
-
-	for (dbnum = 0; dbnum < old_cluster.dbarr.ndbs; dbnum++)
-	{
-		DbInfo	   *olddb = &old_cluster.dbarr.dbs[dbnum];
-		PGconn	   *conn = connectToServer(&new_cluster, olddb->db_name);
-
-		/*
-		 * GPDB doesn't allow hacking the catalogs without setting
-		 * allow_system_table_mods first.
-		 */
-		PQclear(executeQueryOrDie(conn, "set allow_system_table_mods='dml'"));
-
-		/*
-		 * check_mode doesn't do much interesting for this but at least
-		 * we'll know we are allowed to change allow_system_table_mods
-		 * which is required
-		 */
-		if (!check_mode)
-		{
-			PQclear(executeQueryOrDie(conn,
-									  "UPDATE pg_index SET indisvalid = false FROM pg_class c WHERE c.oid = indexrelid AND indexrelid >= %u AND relam = 3013;",
-									  FirstNormalObjectId));
-		}
-		PQfinish(conn);
-	}
-
-	check_ok();
-}
