+++ contrib/pg_upgrade/version_old_8_3.c
@@ -54,8 +54,8 @@ old_8_3_check_for_name_data_type_usage(ClusterInfo *cluster)
 								"FROM	pg_catalog.pg_class c, "
 								"		pg_catalog.pg_namespace n, "
 								"		pg_catalog.pg_attribute a "
-								"WHERE	c.relkind='r' AND "
-								"		c.oid = a.attrelid AND "
+								"WHERE	c.oid = a.attrelid AND "
+								"		c.relkind='r' AND "
 								"		a.attnum > 1 AND "
 								"		NOT a.attisdropped AND "
 								"		a.atttypid = 'pg_catalog.name'::pg_catalog.regtype AND "
